--- a/.github/workflows/main.yml
+++ b/.github/workflows/main.yml
@@ -8,7 +8,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, windows-latest]
-        python-version: [ '3.7', '3.8', '3.9', '3.10', '3.11', 'pypy3.9' ]
+        python-version: [ '3.8', '3.9', '3.10', '3.11', '3.12', 'pypy3.9' ]
         exclude:
           - os: windows-latest
             python-version: pypy3.9