--- a/construct/lib/containers.py
+++ b/construct/lib/containers.py
@@ -1,11 +1,6 @@
 from construct.lib.py3compat import *
 import re
-import collections
 import sys
-
-OrderedDict = dict
-if sys.version_info < (3, 7):
-    OrderedDict = collections.OrderedDict
 
 
 globalPrintFullStrings = False
@@ -61,9 +56,9 @@ def recursion_lock(retval="<recursion detected>", lock_name="__recursion_lock__"
     return decorator
 
 
-class Container(OrderedDict):
+class Container(dict):
     r"""
-    Generic ordered dictionary that allows both key and attribute access, and preserves key order by insertion. Adding keys is preferred using \*\*entrieskw (requires Python 3.6). Equality does NOT check item order. Also provides regex searching.
+    Generic ordered dictionary that allows both key and attribute access, and preserves key order by insertion. Adding keys is preferred using \*\*entrieskw. Equality does NOT check item order. Also provides regex searching.
 
     Note that not all parameters can be accessed via attribute access (dot operator). If the name of an item matches a method name of the Container, it can only be accessed via key acces (square brackets). This includes the following names: clear, copy, fromkeys, get, items, keys, move_to_end, pop, popitem, search, search_all, setdefault, update, values.
 
@@ -71,9 +66,8 @@ class Container(OrderedDict):
 
         # empty dict
         >>> Container()
-        # list of pairs, not recommended
-        >>> Container([ ("name","anonymous"), ("age",21) ])
-        # This syntax requires Python 3.6
+        # sequence of pairs
+        >>> Container([("name", "anonymous"), ("age", 21)])
         >>> Container(name="anonymous", age=21)
         # copies another dict
         >>> Container(dict2)
@@ -88,71 +82,48 @@ class Container(OrderedDict):
             text = u'utf8 decoded string...' (total 22)
             value = 123
     """
-    __slots__ = ["__recursion_lock__"]
+    __slots__ = ('__dict__', '__recursion_lock__')
 
-    def __getattr__(self, name):
-        try:
-            if name in self.__slots__:
-                return object.__getattribute__(self, name)
-            else:
-                return self[name]
-        except KeyError:
-            raise AttributeError(name)
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.__dict__ = self
 
-    def __setattr__(self, name, value):
-        try:
-            if name in self.__slots__:
-                return object.__setattr__(self, name, value)
-            else:
-                self[name] = value
-        except KeyError:
-            raise AttributeError(name)
+    def copy(self, /):
+        return self.__class__(self)
 
-    def __delattr__(self, name):
-        try:
-            if name in self.__slots__:
-                return object.__delattr__(self, name)
-            else:
-                del self[name]
-        except KeyError:
-            raise AttributeError(name)
+    def __copy__(self, /):
+        return self.copy()
 
-    def update(self, seqordict):
-        """Appends items from another dict/Container or list-of-tuples."""
-        if isinstance(seqordict, dict):
-            seqordict = seqordict.items()
-        for k,v in seqordict:
-            self[k] = v
+    # this is required because otherwise deepcopy will
+    # copy self and self.__dict__ separately for some reason
+    def __deepcopy__(self, _, /):
+        return self.copy()
 
-    def copy(self):
-        return Container(self)
-
-    __update__ = update
-
-    __copy__ = copy
-
-    def __dir__(self):
+    def __dir__(self, /):
         """For auto completion of attributes based on container values."""
         return list(self.keys()) + list(self.__class__.__dict__) + dir(super(Container, self))
 
-    def __eq__(self, other):
+    def __eq__(self, other, /):
         if self is other:
             return True
         if not isinstance(other, dict):
             return False
+
         def isequal(v1, v2):
             if v1.__class__.__name__ == "ndarray" or v2.__class__.__name__ == "ndarray":
                 import numpy
+
                 return numpy.array_equal(v1, v2)
             return v1 == v2
-        for k,v in self.items():
+
+        for k, v in self.items():
             if isinstance(k, unicodestringtype) and k.startswith(u"_"):
                 continue
             if isinstance(k, bytestringtype) and k.startswith(b"_"):
                 continue
             if k not in other or not isequal(v, other[k]):
                 return False
-        for k,v in other.items():
+        for k, v in other.items():
             if isinstance(k, unicodestringtype) and k.startswith(u"_"):
                 continue
             if isinstance(k, bytestringtype) and k.startswith(b"_"):
@@ -161,13 +132,13 @@ class Container(OrderedDict):
                 return False
         return True
 
-    def __ne__(self, other):
-       return not self == other
+    def __ne__(self, other, /):
+        return not self == other
 
     @recursion_lock()
-    def __repr__(self):
+    def __repr__(self, /):
         parts = []
-        for k,v in self.items():
+        for k, v in self.items():
             if isinstance(k, str) and k.startswith("_"):
                 continue
             if isinstance(v, stringtypes):
@@ -177,20 +148,26 @@ class Container(OrderedDict):
         return "Container(%s)" % ", ".join(parts)
 
     @recursion_lock()
-    def __str__(self):
+    def __str__(self, /):
         indentation = "\n    "
         text = ["Container: "]
         isflags = getattr(self, "_flagsenum", False)
-        for k,v in self.items():
+        for k, v in self.items():
             if isinstance(k, str) and k.startswith("_") and not globalPrintPrivateEntries:
                 continue
             if isflags and not v and not globalPrintFalseFlags:
                 continue
             text.extend([indentation, str(k), " = "])
             if v.__class__.__name__ == "EnumInteger":
-                text.append("(enum) (unknown) %s" % (v, ))
+                text.append("(enum) (unknown) %s" % (v,))
             elif v.__class__.__name__ == "EnumIntegerString":
-                text.append("(enum) %s %s" % (v, v.intvalue, ))
+                text.append(
+                    "(enum) %s %s"
+                    % (
+                        v,
+                        v.intvalue,
+                    )
+                )
             elif v.__class__.__name__ in ["HexDisplayedBytes", "HexDumpDisplayedBytes"]:
                 text.append(indentation.join(str(v).split("\n")))
             elif isinstance(v, bytestringtype):
@@ -198,22 +175,26 @@ class Container(OrderedDict):
                 if len(v) <= printingcap or globalPrintFullStrings:
                     text.append("%s (total %d)" % (reprstring(v), len(v)))
                 else:
-                    text.append("%s... (truncated, total %d)" % (reprstring(v[:printingcap]), len(v)))
+                    text.append(
+                        "%s... (truncated, total %d)" % (reprstring(v[:printingcap]), len(v))
+                    )
             elif isinstance(v, unicodestringtype):
                 printingcap = 32
                 if len(v) <= printingcap or globalPrintFullStrings:
                     text.append("%s (total %d)" % (reprstring(v), len(v)))
                 else:
-                    text.append("%s... (truncated, total %d)" % (reprstring(v[:printingcap]), len(v)))
+                    text.append(
+                        "%s... (truncated, total %d)" % (reprstring(v[:printingcap]), len(v))
+                    )
             else:
                 text.append(indentation.join(str(v).split("\n")))
         return "".join(text)
 
-    def _search(self, compiled_pattern, search_all):
+    def _search(self, compiled_pattern, search_all, /):
         items = []
         for key in self.keys():
             try:
-                if isinstance(self[key], (Container,ListContainer)):
+                if isinstance(self[key], (Container, ListContainer)):
                     ret = self[key]._search(compiled_pattern, search_all)
                     if ret is not None:
                         if search_all:
@@ -246,14 +227,13 @@ class Container(OrderedDict):
         compiled_pattern = re.compile(pattern)
         return self._search(compiled_pattern, True)
 
-    def __getstate__(self):
+    def __getstate__(self, /):
         """
         Used by pickle to serialize an instance to a dict.
         """
-        ret = OrderedDict(self)
-        return ret
+        return dict(self)
 
-    def __setstate__(self, state):
+    def __setstate__(self, state, /):
         """
         Used by pickle to de-serialize from a dict.
         """
@@ -272,8 +252,10 @@ class ListContainer(list):
 
     ::
 
+        >>> obj
+        ListContainer([1, 2, 3])
         >>> print(repr(obj))
-        [1, 2, 3]
+        ListContainer([1, 2, 3])
         >>> print(obj)
         ListContainer
             1
@@ -282,11 +264,11 @@ class ListContainer(list):
     """
 
     @recursion_lock()
-    def __repr__(self):
-        return "ListContainer(%s)" % (list.__repr__(self), )
+    def __repr__(self, /):
+        return "ListContainer(%s)" % (list.__repr__(self),)
 
     @recursion_lock()
-    def __str__(self):
+    def __str__(self, /):
         indentation = "\n    "
         text = ["ListContainer: "]
         for k in self:
@@ -295,7 +277,7 @@ class ListContainer(list):
             text.append(indentation.join(lines))
         return "".join(text)
 
-    def _search(self, compiled_pattern, search_all):
+    def _search(self, compiled_pattern, search_all, /):
         items = []
         for item in self:
             try: