Change `Container` implementation. This implementation is **5x** faster on my machine on CPython3.12.
This is a **backwards incompatible** change:
- **the biggest problem:** shadowing behaviour changed. Before `con.update` was always referring to method, now it will refer to container item called `update` (if it exists, otherwise it will be a method). I added new test to ensure that behaviour. All other container-related tests pass perfectly. Documentation was updated too. There is a couple in the `core.py` that was using methods directly, they probably should be updated to do `Container.method(ctx, ...)`
- `__(get|set|del)attr__` methods were removed. If anybody is calling them directly (`con.__getattr__('foo')`), this code will break
- `Container.update` method was removed. It wasn't doing anything that `dict.update` is not able to do. Nothing should break
- `__update__` method was removed. Not sure what it was doing. It is not referenced anywhere in the codebase, and it is not a special dunder.
- some private and dunder methods don't accept kwargs anymore
- whoops, I accidentally ran `black` on the file, so there is a couple of formatting changes 

Benchmark code:
```py
import timeit
N = 10**8
for _ in range(10):
    t = timeit.timeit('c.x', 'from construct import Container; c = Container(x=5)', number=N) / N
    print(round(t * 10**9), 'ns')

```
Original implementation:
```py
149 ns
164 ns
164 ns
163 ns
163 ns
174 ns
162 ns
165 ns
169 ns
163 ns
```
New implementation:
```py
26 ns
26 ns
28 ns
26 ns
30 ns
30 ns
31 ns
34 ns
31 ns
31 ns
```