@denballakh 2023-12-13
I also ran already written benchmarks.
I ran following commands on master and this branch:
`py.test test_benchmarks.py --benchmark-save=master       --benchmark-only --benchmark-calibration-precision=100 --benchmark-max-time=5`
`py.test test_benchmarks.py --benchmark-save=newcontainer --benchmark-only --benchmark-calibration-precision=100 --benchmark-max-time=5`

Results:
[0004_master.json](https://github.com/construct/construct/files/13655961/0004_master.json)
[0005_newcontainer.json](https://github.com/construct/construct/files/13655963/0005_newcontainer.json)

I am comparing results using this command:
`py.test-benchmark compare 0004 0005 --sort=name --group-by=name --columns="mean, stddev"`
New implementation gives 0-20% increase in different cases.
`overall` benchmark does a lot of different things, I can see 3-4% in perfomance increase:
```py
----------------- benchmark 'test_overall_build': 2 tests ------------------
Name (time in us)                         Mean              StdDev
----------------------------------------------------------------------------
test_overall_build (0004_master)      930.0496 (1.04)      86.0212 (1.0)
test_overall_build (0005_newcont)     893.4758 (1.0)      101.2078 (1.18)
----------------------------------------------------------------------------

----------------- benchmark 'test_overall_build_compiled': 2 tests -----------------
Name (time in us)                                  Mean             StdDev
------------------------------------------------------------------------------------
test_overall_build_compiled (0004_master)      758.1896 (1.03)     83.4523 (1.0)
test_overall_build_compiled (0005_newcont)     734.9712 (1.0)      88.5543 (1.06)
------------------------------------------------------------------------------------

--------------- benchmark 'test_overall_parse': 2 tests ----------------
Name (time in ms)                       Mean            StdDev
------------------------------------------------------------------------
test_overall_parse (0004_master)      1.2821 (1.03)     0.1490 (1.0)
test_overall_parse (0005_newcont)     1.2494 (1.0)      0.1656 (1.11)
------------------------------------------------------------------------

------------------ benchmark 'test_overall_parse_compiled': 2 tests -----------------
Name (time in us)                                  Mean              StdDev
-------------------------------------------------------------------------------------
test_overall_parse_compiled (0004_master)      988.7750 (1.03)     105.2070 (1.0)
test_overall_parse_compiled (0005_newcont)     962.7802 (1.0)      108.1083 (1.03)
-------------------------------------------------------------------------------------
```