@arekbulski 2023-12-13
Wow these results look great. Good job. Do not worry about breaking changes. 

Do NOT update the docs yet. I am doing heavy work on those. I will let you know when I am done with it.

CI broke, please look into it.