@denballakh 2023-12-13
I undid some changes: docs and some formatting changes.
Errors in CI were caused by `"items"` Struct field in some cases, which broke `con.items()` behaviour.
Added more tests.
The only tests that are failing in my setup are cryptography- and ksy- related. I dont think they should be affected too much by these changes.

A bit more about CI:
I started using pos-only arguments, which were introduced in 3.8.
I suggest dropping support for 3.7 and adding support for 3.12 (I am testing locally on this version, it works fine).
I will do that in next commit