@arekbulski 2023-12-13
Looks like CI is still broken on 3.8.