@denballakh 2023-12-13
It looks like `assert Container.search(c, 'y') == None` line causes python crash. I can reproduce it locally. Investigating...