@denballakh 2023-12-13
Hmmmmm...
If I add simple `print('foo')` in the `Container._search`, bug disappears... If I remove it - it crashes.

I guess it is happening because of recursion (it is testing on recursive container). This is not a realistic situation, so I will make this a normal container.
My guess is that `RecursionError` is being raised, but then it is ignored by `except Exception: pass`, so stack keeps growing, `RecursionError`s are being raised and so on. Eventually it runs out of space for C stack and crashes.

Still have no idea why simple `print('foo')` can fix the problem - probably because of some `pytest` trickery.