@denballakh 2023-12-13
Doing `except: pass` or `except Exception: pass` is not a great idea. These constructs swallow all exceptions, which might be not expected, so possible errors are hidden.

This ^ is an example. All such constructs should be replaced with explicit `except SomeExpectedExceptionThatShouldBeIgnored: pass` eventualy.