@arekbulski 2023-12-14
Okay... I am too tired to work on this today. I will get this PR accepted by tomorrow evening.