@arekbulski 2023-12-14
BTW, the right way to measure timeit is to take a minimum rather than average, or print several measurements:
```py
In [19]: min([ timeit.timeit("f()", "def f(): pass", number=10000000) for i in range(10) ])
Out[19]: 0.26637502499943366
```